requirements.txt
tikzplotlib==0.10.1
jpeg4py==0.1.4
lmdb==1.4.1
pandas==2.1.4
scipy==1.11.4
loguru==0.7.2
kornia==0.7.2
einops==0.7.0
timm==0.9.16
gradio==4.13.0
pillow==10.1.0
albumentations==1.4.1
wandb==0.16.1
progressbar2
gdown >= 4.7.1
gitpython >= 3.1
git+https://github.com/cheind/py-thin-plate-spline
hickle >= 5.0
tensorboard >= 2.11
numpy >= 1.21
git+https://github.com/facebookresearch/segment-anything.git
fastapi==0.111.0
pydantic==2.7.1
opencv-python >= 4.8
matplotlib
pyyaml
av >= 0.5.2
openmim
tqdm >= 4.66.1
psutil
ffmpeg-python
cython
pycocotools >= 2.0.7
hydra-core >= 1.3.2
PySide6 >= 6.2.0
charset-normalizer >= 3.1.0
netifaces >= 0.11.0
cchardet >= 2.1.7
easydict
requests
pyqtdarktheme
imageio == 2.25.0
imageio[ffmpeg]
ffmpeg-python